backend/requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
ez-setup==0.9
gunicorn==19.9.0
psycopg2==2.8.3
python-dotenv==0.10.3
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
